django/requirements.txt
django
psycopg2

Pillow
